implemented addNode, addEdge and addEdgeAndNode in class GraphModel; changed all references to java.awt.Point to java.awt.geom.Point2D.Float so scaling will work; stores max ids for states, transitions and events (since FSA will not be rebuilt only on save  but will instead be subjected to deletions and additions of elements by multiple views at runtime we cannot use storage indices as unique ids).

+++ src/model/fsa/ver1/State.java
@@ -1,13 +1,9 @@
 package model.fsa.ver1;
 
-import io.fsa.ver1.SubElement;
 import io.fsa.ver1.SubElementContainer;
 
-import java.awt.Point;
 import java.util.LinkedList;
 import java.util.ListIterator;
-
-import presentation.fsa.StateLayout;
 
 import model.fsa.FSATransition;
 
@@ -31,7 +27,7 @@ public class State extends SubElementContainer implements model.fsa.FSAState {
      * constructs a state with the given id.
      * @param id the id of the state.
      */
-    public State(int id){
+    public State(long id){
         this.id = id;
         sourceT = new LinkedList<FSATransition>();
         targetT = new LinkedList<FSATransition>();
